Combined Aeon and Ill links jest tests to create a SpecialRequestLink test
